+++ app/src/main/AndroidManifest.xml
@@ -13,12 +13,12 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.CoinCap"
+        android:theme="@style/CoinCap.Theme"
         tools:targetApi="31">
         <activity
             android:name=".MainActivity"
             android:exported="true"
-            android:theme="@style/Theme.CoinCap">
+            android:theme="@style/CoinCap.Theme">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
