+++ core/domain/src/main/java/com/client/domain/usecase/home_screen/rates/GetRatesUseCase.kt
@@ -1,4 +1,4 @@
-package com.client.domain.usecase.home_screen
+package com.client.domain.usecase.home_screen.rates
 
 import com.client.data.model.Rate
 import com.client.data.network.Result
